chore: move weather.js to its own folder instead of test folder
